+++ src/pages/LinePage/LinePage.stories.tsx
@@ -1,13 +1,12 @@
 import LinePage from './LinePage';
 import { Story } from '@storybook/react';
-import { PageProps } from '../types';
 
 export default {
   title: 'pages/LinePage',
   component: LinePage,
 };
 
-const Template: Story<PageProps> = (args) => <LinePage {...args} />;
+const Template: Story = (args) => <LinePage {...args} />;
 
 export const Default = Template.bind({});
