+++ src/pages/LoginPage/LoginPage.stories.tsx
@@ -1,13 +1,12 @@
 import LoginPage from './LoginPage';
 import { Story } from '@storybook/react';
-import { PageProps } from '../types';
 
 export default {
   title: 'pages/LoginPage',
   component: LoginPage,
 };
 
-const Template: Story<PageProps> = (args) => <LoginPage {...args} />;
+const Template: Story = (args) => <LoginPage {...args} />;
 
 export const Default = Template.bind({});
