+++ src/pages/SectionPage/SectionPage.stories.tsx
@@ -1,13 +1,12 @@
 import StationPage from './SectionPage';
 import { Story } from '@storybook/react';
-import { PageProps } from '../types';
 
 export default {
   title: 'pages/StationPage',
   component: StationPage,
 };
 
-const Template: Story<PageProps> = (args) => <StationPage {...args} />;
+const Template: Story = (args) => <StationPage {...args} />;
 
 export const Default = Template.bind({});
