+++ src/pages/SignupPage/SignupPage.stories.tsx
@@ -1,13 +1,12 @@
 import SignupPage from './SignupPage';
 import { Story } from '@storybook/react';
-import { PageProps } from '../types';
 
 export default {
   title: 'pages/SignupPage',
   component: SignupPage,
 };
 
-const Template: Story<PageProps> = (args) => <SignupPage {...args} />;
+const Template: Story = (args) => <SignupPage {...args} />;
 
 export const Default = Template.bind({});
